02-workflow-orchestration/homework/requirements/base.txt
apache-airflow-providers-google
pyarrow 
# TO-DO: remove this and place it toask or dag specific requirmenet with virtual env